Provide separate calendar page for admins.
