+++ src/components/pages/CalendarPage/AdminCalendarPage/index.js
@@ -1,0 +1,3 @@
+import AdminCalendarPage from './AdminCalendarPage';
+
+export default AdminCalendarPage;
