+++ src/components/pages/CalendarPage/VolunteerCalendarPage/index.js
@@ -1,0 +1,3 @@
+import VolunteerCalendarPage from './VolunteerCalendarPage';
+
+export default VolunteerCalendarPage;
